Hide unsupported integer suffixes from rustc 1.15
